When champion.gg is offline , _PatchNumber is empty and an update is triggered , this should fix the problem

+++ LoLMasteryManager/MasteryManager.vb
@@ -68,7 +68,7 @@ Public Class MasteryManager
 
         If oMetadata IsNot Nothing Then
 
-            If Not String.Equals(_PatchNumber, oMetadata.PatchNumber) Then
+            If Not String.Equals(_PatchNumber, oMetadata.PatchNumber) AndAlso Not String.Equals(_PatchNumber, "") Then
 
                 Directory.Delete(Directories.Data, True)
